# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pâtisserie ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: List categories
      responses:
        "200":
          description: Categories list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      summary: Update category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryBody"
      responses:
        "200":
          description: Category updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  # UNITS
  /units:
    get:
      operationId: getUnits
      tags: [units]
      summary: List units
      responses:
        "200":
          description: Units list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Unit"
    post:
      operationId: createUnit
      tags: [units]
      summary: Create unit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUnitBody"
      responses:
        "201":
          description: Unit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"

  # PRODUCTS
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: List products
      parameters:
        - name: type
          in: query
          schema:
            type: ["string", "null"]
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: lowStock
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      summary: Get product
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      summary: Update product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}/stock:
    get:
      operationId: getProductStock
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      summary: Get product stock by branch
      responses:
        "200":
          description: Stock by branch
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockLevel"

  # STOCK
...
components:
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        parentId:
          type: ["integer", "null"]
        parentName:
          type: ["string", "null"]
        productCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - productCount
        - createdAt
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        parentId:
          type: ["integer", "null"]
      required:
        - name
    UpdateCategoryBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        parentId:
          type: ["integer", "null"]
    Unit:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        abbreviation:
          type: string
        allowDecimals:
          type: boolean
      required:
        - id
        - name
        - abbreviation
        - allowDecimals
    CreateUnitBody:
      type: object
      properties:
        name:
          type: string
        abbreviation:
          type: string
        allowDecimals:
          type: boolean
      required:
        - name
        - abbreviation
        - allowDecimals
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sku:
          type: ["string", "null"]
        barcode:
          type: ["string", "null"]
        type:
          type: string
          enum: [finished, ingredient, packaging, semi_finished]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        unitId:
          type: integer
        unitName:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        costPrice:
          type: number
        sellingPrice:
          type: number
        alertQuantity:
          type: ["number", "null"]
        shelfLifeDays:
          type: ["integer", "null"]
        isManaged:
          type: boolean
        isSellable:
          type: boolean
        isPurchasable:
          type: boolean
        isFabricated:
          type: boolean
        totalStock:
          type: number
        createdAt:
          type: string
      required:
        - id
        - name
        - type
        - unitId
        - unitName
        - costPrice
        - sellingPrice
        - isManaged
        - isSellable
        - isPurchasable
        - isFabricated
        - totalStock
        - createdAt
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        sku:
          type: ["string", "null"]
        barcode:
          type: ["string", "null"]
        type:
          type: string
          enum: [finished, ingredient, packaging, semi_finished]
        categoryId:
          type: ["integer", "null"]
        unitId:
          type: integer
        description:
          type: ["string", "null"]
        costPrice:
          type: number
        sellingPrice:
          type: number
        alertQuantity:
          type: ["number", "null"]
        shelfLifeDays:
          type: ["integer", "null"]
        isManaged:
          type: boolean
        isSellable:
          type: boolean
        isPurchasable:
          type: boolean
        isFabricated:
          type: boolean
      required:
        - name
        - type
        - unitId
        - costPrice
        - sellingPrice
        - isManaged
        - isSellable
        - isPurchasable
        - isFabricated
    UpdateProductBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
        barcode:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        unitId:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
        costPrice:
          type: ["number", "null"]
        sellingPrice:
          type: ["number", "null"]
        alertQuantity:
          type: ["number", "null"]
        isManaged:
          type: ["boolean", "null"]
        isSellable:
          type: ["boolean", "null"]
        isPurchasable:
          type: ["boolean", "null"]
        isFabricated:
          type: ["boolean", "null"]
    StockLevel:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        productType:
          type: string
        branchId:
          type: integer
        branchName:
          type: string
        quantity:
          type: number
        alertQuantity:
          type: ["number", "null"]
        unitName:
          type: string
        status:
          type: string
          enum: [ok, low, critical, out]
        valueCost:
          type: number
      required:
        - productId
        - productName
        - productType
        - branchId
        - branchName
        - quantity
        - unitName
        - status
        - valueCost
...